svc/server/requirements.txt
fastapi[standard]
pydantic
aiosqlite
dotenv
typing-extensions
google-genai[aiohttp]